% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R, R/spop_cycle_raw.R, R/spop_no_cycle.R
\docType{methods}
\name{add<-}
\alias{add<-}
\alias{add<-,spop,data.frame-method}
\alias{add<-,spop-method}
\alias{add<-}
\alias{add<-,spop,data.frame-method}
\alias{add<-,spop-method}
\alias{add<-}
\alias{add<-,spop,data.frame-method}
\alias{add<-,spop-method}
\title{Add batch}
\usage{
add(x) <- value

\S4method{add}{spop,data.frame}(x) <- value

add(x) <- value

\S4method{add}{spop,data.frame}(x) <- value

add(x) <- value

\S4method{add}{spop,data.frame}(x) <- value
}
\arguments{
\item{x}{spop class instant}

\item{value}{\code{data.frame} with \code{name} and \code{age} fields}

\item{x}{spop class instant}

\item{value}{\code{data.frame} with \code{name} and \code{age} fields}

\item{x}{spop class instant}

\item{value}{\code{data.frame} with \code{name}, \code{age}, and \code{development} fields}
}
\description{
Introduce a batch of individuals with a given age

Introduce a batch of individuals with a given age

Introduce a batch of individuals with a given age and development stage. By default, development stage is set the same as age
}
