% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albopictus.R
\docType{package}
\name{albopictus-package}
\alias{albopictus}
\alias{albopictus-package}
\title{Age-Structured Population Dynamics Model}
\description{
Implements the "spop" class for age-structured population dynamics modelling.
For more information, see class description.
}
\references{
\itemize{
\item Kamil Erguler, Stephanie E Smith-Unna, Joanna Waldock, Yiannis Proestos, George K Christophides, Jos Lelieveld, Paul E Parham. Large-scale modelling of the environmentally-driven population dynamics of temperate Aedes albopictus (Skuse). PLOS ONE, 2016
\item Kamil Erguler, Nastassya L Chandra, Yiannis Proestos, Jos Lelieveld, George K Christophides, Paul E Parham. A large-scale stochastic spatiotemporal model for Aedes albopictus-borne chikungunya epidemiology. PLOS ONE, 2017
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kerguler/albopictusR}
}

}
\author{
Kamil Erguler, Associate Research Scientist, EEWRC, The Cyprus Institute \email{k.erguler@cyi.ac.cy}
}
