\name{albopictus}
\alias{albopictus}
\title{ Age-structured population dynamics model }
\description{
This package implements the "spop" class for age-structured population dynamics modelling. For more information, see the class description.
}
\author{
\itemize{
    \item Kamil Erguler, Associate Research Scientist, EEWRC, The Cyprus Institute \email{k.erguler@cyi.ac.cy}
    }
}
