% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R, R/spop_cycle_raw.R
\docType{methods}
\name{completed}
\alias{completed}
\alias{completed,spop-method}
\alias{completed}
\alias{completed,spop-method}
\title{Read completed}
\usage{
completed(x)

\S4method{completed}{spop}(x)

completed(x)

\S4method{completed}{spop}(x)
}
\arguments{
\item{x}{spop class instant}

\item{x}{spop class instant}
}
\description{
Read the number of individuals completing a period after each iteration

Read the number of individuals completing a period after each iteration
}
