% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{devtable}
\alias{devtable}
\alias{devtable,spop-method}
\title{Read devtable}
\usage{
devtable(x)

\S4method{devtable}{spop}(x)
}
\arguments{
\item{x}{spop class instant}
}
\description{
Read the number, age, and development cycles of individuals completing their development after each iteration
}
