% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R, R/spop_cycle_raw.R, R/spop_no_cycle.R
\docType{methods}
\name{iterate<-}
\alias{iterate<-}
\alias{iterate<-,spop,data.frame-method}
\alias{iterate<-,spop-method}
\alias{iterate<-}
\alias{iterate<-,spop,data.frame-method}
\alias{iterate<-,spop-method}
\alias{iterate<-}
\alias{iterate<-,spop,data.frame-method}
\alias{iterate<-,spop-method}
\title{Iterate population}
\usage{
iterate(x) <- value

\S4method{iterate}{spop,data.frame}(x) <- value

iterate(x) <- value

\S4method{iterate}{spop,data.frame}(x) <- value

iterate(x) <- value

\S4method{iterate}{spop,data.frame}(x) <- value
}
\arguments{
\item{x}{spop class instant}

\item{value}{\code{data.frame} with the following fields
\itemize{
\item \code{dev}
\item \code{dev_mean}, \code{dev_sd}
\item \code{death}
\item \code{death_mean}, \code{death_sd}
}}

\item{x}{spop class instant}

\item{value}{\code{data.frame} with the following fields
\itemize{
\item \code{dev}
\item \code{dev_mean}, \code{dev_sd}
\item \code{death}
\item \code{death_mean}, \code{death_sd}
}}

\item{x}{spop class instant}

\item{value}{\code{data.frame} with the following fields
\itemize{
\item \code{dev}
\item \code{dev_mean}, \code{dev_sd}
\item \code{death}
\item \code{death_mean}, \code{death_sd}
}}
}
\description{
Iterate the population for one day

Iterate the population for one day

Iterate the population for one day
}
