% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\name{nbinom_dist_prob}
\alias{nbinom_dist_prob}
\title{Negative binomial probability of death or development}
\usage{
nbinom_dist_prob(xr, mn, std)
}
\arguments{
\item{xr}{age of individuals}

\item{mn}{mean age of death or development}

\item{std}{standard deviation of the age of death or development}
}
\description{
Negative binomial-distributed probability of death or development happening in the next iteration
}
