% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{perturbate<-}
\alias{perturbate<-}
\alias{perturbate<-,spop,data.frame-method}
\alias{perturbate<-,spop-method}
\title{Iterate population without incrementing age or development}
\usage{
perturbate(x) <- value

\S4method{perturbate}{spop,data.frame}(x) <- value
}
\arguments{
\item{x}{spop class instant}

\item{value}{\code{data.frame} with the following fields
\itemize{
\item \code{death}: age-independent daily probability of death
\item \code{death_mean} and \code{death_sd}: age-dependent daily probability of death (\code{death_sd=0} indicates fixed life time (defined by \code{death_mean}))
\item \code{dev}: age-independent daily probability of development
\item \code{dev_mean} and \code{dev_sd}: age-dependent daily probability of development (\code{dev_sd=0} indicates fixed development time (defined by \code{dev_mean}))
}}
}
\description{
Iterate the population for one day keeping age and development fixed
}
