% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R, R/spop_cycle_raw.R, R/spop_no_cycle.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,spop-method}
\alias{size}
\alias{size,spop-method}
\alias{size}
\alias{size,spop-method}
\title{Read size}
\usage{
size(x)

\S4method{size}{spop}(x)

size(x)

\S4method{size}{spop}(x)

size(x)

\S4method{size}{spop}(x)
}
\arguments{
\item{x}{spop class instant}

\item{x}{spop class instant}

\item{x}{spop class instant}
}
\description{
Read the total number of individuals

Read the total number of individuals

Read the total number of individuals
}
