% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{add<-}
\alias{add<-}
\alias{add<-,spop,data.frame-method}
\alias{add<-,spop-method}
\title{Add batch}
\usage{
add(x) <- value

\S4method{add}{spop,data.frame}(x) <- value
}
\arguments{
\item{x}{spop class instant}

\item{value}{\code{data.frame} with \code{name} and \code{age} fields}
}
\description{
Introduce a batch of individuals with a given age
}
