\name{albopictus}
\alias{albopictus}
\title{ Age-Structured Population Dynamics Model }
\description{
Implements the "spop" class for age-structured population dynamics modelling. For more information, see class description.
}
\author{
\itemize{
    \item Kamil Erguler, Associate Research Scientist, EEWRC, The Cyprus Institute \email{k.erguler@cyi.ac.cy}
    }
}
