% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{dead}
\alias{dead}
\alias{dead,spop-method}
\title{Read dead}
\usage{
dead(x)

\S4method{dead}{spop}(x)
}
\arguments{
\item{x}{spop class instant}
}
\description{
Read the number of dead individuals after each iteration
}
