% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{developed}
\alias{developed}
\alias{developed,spop-method}
\title{Read developed}
\usage{
developed(x)

\S4method{developed}{spop}(x)
}
\arguments{
\item{x}{spop class instant}
}
\description{
Read the number of individuals designated to complete their development
}
