% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spop.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,spop-method}
\title{Read size}
\usage{
size(x)

\S4method{size}{spop}(x)
}
\arguments{
\item{x}{spop class instant}
}
\description{
Read the total number of individuals
}
